Move http/socks proxy stuff to separate file
(cherry picked from commit 13e911378764cafb98030ebe80832739ae5ce87e)
